Player Jump Mechanic implemented!
Player can jump with Swipe Up!

//TODO:
We need a mechanism that 2 or more players can play and win. There would be some stars to collect and increases your score. At the end of the game, the winner will be chosen with its score...

DOTween package added.
ProBuilder Unity package added.